requirements.txt
flask
pillow
easyocr
openai
